+++ apps/etterlatte-saksbehandling-ui/client/src/components/person/OpprettNyRevurdering.tsx
@@ -3,10 +3,10 @@ import React, { useState } from 'react'
 import { useApiCall } from '~shared/hooks/useApiCall'
 import { Revurderingaarsak, tekstRevurderingsaarsak } from '~shared/types/Revurderingaarsak'
 import { useNavigate } from 'react-router-dom'
-import { ButtonGroup } from '~components/person/VurderHendelseModal'
 import { opprettRevurdering as opprettRevurderingApi } from '~shared/api/revurdering'
 
 import { isPending } from '~shared/api/apiUtils'
+import { ButtonGroup } from '~shared/styled'
 
 export const OpprettNyRevurdering = ({
   sakId,
