+++ apps/etterlatte-saksbehandling-ui/client/src/components/person/hendelser/ArkivertHendelseExpandableRow.tsx
@@ -3,7 +3,7 @@ import { Table } from '@navikt/ds-react'
 import { HendelseBeskrivelse } from '~components/person/hendelser/HendelseBeskrivelse'
 import { SakType } from '~shared/types/sak'
 import { Grunnlagsendringshendelse } from '~components/person/typer'
-import { grunnlagsendringsTittel } from '~components/person/uhaandtereHendelser/utils'
+import { grunnlagsendringsTittel } from '~components/person/hendelser/utils'
 import { formaterStringDato } from '~utils/formattering'
 
 interface Props {
