+++ apps/etterlatte-saksbehandling-ui/client/src/components/person/hendelser/InstitusjonsoppholdVurderingBegrunnelse.tsx
